Rename cursors and components modules
