+++ docs/api/component.rst
@@ -1,0 +1,5 @@
+phasorpy.component
+------------------
+
+.. automodule:: phasorpy.component
+    :members:
