--- docs/api/components.rst
@@ -1,5 +1,0 @@
-phasorpy.components
--------------------
-
-.. automodule:: phasorpy.components
-    :members:
