+++ docs/api/cursor.rst
@@ -1,0 +1,5 @@
+phasorpy.cursor
+---------------
+
+.. automodule:: phasorpy.cursor
+    :members:
