--- docs/api/cursors.rst
@@ -1,5 +1,0 @@
-phasorpy.cursors
-----------------
-
-.. automodule:: phasorpy.cursors
-    :members:
