+++ docs/conf.py
@@ -138,11 +138,11 @@ class TutorialOrder:
         # api
         'io',
         'phasor_from_lifetime',
-        'multi-harmonic',
-        'filtering',
+        'multi_harmonic',
+        'filter',
         'phasorplot',
-        'cursors',
-        'components',
+        'cursor',
+        'component',
         'fret',
         'lifetime_to_signal',
         'pca',
